#3 pages에 feed 관련 page 정리
